+++ src/pages/Processos/ImportacaoDados/Vagas/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { screen, fireEvent } from '@testing-library/react';
 import { App } from 'antd';
-import Vagas from '../index';
+import Vagas from '../VagasTela';
 import { renderWithProviders } from '../../../../../test-utils';
 import { useImportacaoDadosVagas } from '../hooks/useImportacaoDadosVagas';
 import { useCargos } from '../../../../../hooks/useCargos';
